requirements.txt
Cython==0.24
numpy==1.10.4
pandas==0.18.1
pandas-datareader==0.2.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
requests-file==1.4
scikit-learn==0.17.1
scipy==0.17.1
TA-Lib==0.4.9
pyyaml==3.11
requests_mock==1.0.0
pytest==3.0.5
pytest_mock==1.2
tensorflow==1.0.0
